Adding task definitions for ECS
The example system was moved to ECS.  I've copied in the task-defintion json for each service.  These would be needed in order to recreate the example system on ECS

I also deleted the docker-compose file, as these are no longer run locally using this method

I also moved some folders under the example-system because they really are part of it.