Updating docker image data to support a second server
I was still seeing issues with things like my logging interfering with the example app.  This wouldn't happen to such an extent in a real system, because there would be multiple servers.  I now have two servers, using AWS elastic IPs so that their IP addresses are static

One server is for the example app
One server is for addons for that app
  * InfluxDB server for collecting logs
  * Grafana for visually seeing the log data
  * Traffic simulator for sending requests

I also trimmed down the number of Elasticsearch nodes to a minimum.  Having a production-like cluster was overkill and requires too many resources.  Going down to a smaller cluster doesn't hurt this project proof of concept.