Adding generate report as a url endpoint
Right when the alert goes off, the report likely isn't quite ready.  Now, you can go to a url to have it generated, or regenerated.  The process of generating it only takes a few seconds.
This requires a jar file from clover in order to work, but I don't want to track it in git.  It'll likely end up in S3, but I'm just leaving it untracked for now.